requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
redis>=4.3.4
neo4j>=5.3.0
python-dotenv>=0.19.0
pytest>=7.0.0
spacy>=3.7.0
transformers>=4.30.0
gensim>=4.3.0
pydantic-settings>=2.0.0
python-multipart
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-3.7.0/es_core_news_md-3.7.0-py3-none-any.whl